Set default concurrency if options are not passed.
The addition of concurrency options breaks many modules that already rely on
using tilelive-streaming without passing concurrency settings. This can be
avoided by using the default concurrency settings if no options are passed in.
Further, an undeclared CONCURRENCY variable would cause the program to crash.
This can also be avoided by declaring this value and setting it to the default
if no options are given.